- Add build based on nuke - Migrate tests to xunit
