+++ 01_patch.diff
@@ -1,0 +1,87 @@
+From 0423c52445fdbcc599c9f794abedcde1451bb4ee Mon Sep 17 00:00:00 2001
+From: ITDancer139 <kevinitdancersommer@gmail.com>
+Date: Sat, 24 Apr 2021 10:32:55 +0200
+Subject: [PATCH] Upgrade dependencies to latest version of .NET Core 3.1
+
+---
+ Sieve/Sieve.csproj                   | 8 ++++----
+ SieveTests/SieveTests.csproj         | 7 ++++---
+ SieveUnitTests/SieveUnitTests.csproj | 3 ++-
+ 3 files changed, 10 insertions(+), 8 deletions(-)
+
+diff --git a/Sieve/Sieve.csproj b/Sieve/Sieve.csproj
+index 344c5af..241c19c 100644
+--- a/Sieve/Sieve.csproj
++++ b/Sieve/Sieve.csproj
+@@ -1,7 +1,7 @@
+ ﻿<Project Sdk="Microsoft.NET.Sdk">
+ 
+   <PropertyGroup>
+-    <TargetFramework>netstandard2.0</TargetFramework>
++    <TargetFramework>netstandard2.1</TargetFramework>
+     <Version>2.3.3</Version>
+     <Description>Sieve is a simple, clean, and extensible framework for .NET Core that adds sorting, filtering, and pagination functionality out of the box. Most common use case would be for serving ASP.NET Core GET queries. Documentation available on GitHub: https://github.com/Biarity/Sieve/
+ </Description>
+@@ -20,9 +20,9 @@ Don't process when filterTerm.Values is null (#59)
+   </PropertyGroup>
+ 
+   <ItemGroup>
+-    <PackageReference Include="Microsoft.CSharp" Version="4.4.1" />
+-    <PackageReference Include="Microsoft.Extensions.Options" Version="2.0.0" />
+-    <PackageReference Include="System.ComponentModel.Annotations" Version="4.4.1" />
++    <PackageReference Include="Microsoft.CSharp" Version="4.7.0" />
++    <PackageReference Include="Microsoft.Extensions.Options" Version="3.1.14" />
++    <PackageReference Include="System.ComponentModel.Annotations" Version="5.0.0" />
+   </ItemGroup>
+ 
+ </Project>
+diff --git a/SieveTests/SieveTests.csproj b/SieveTests/SieveTests.csproj
+index 1374e21..4d4cd8c 100644
+--- a/SieveTests/SieveTests.csproj
++++ b/SieveTests/SieveTests.csproj
+@@ -1,7 +1,7 @@
+ ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
+ 
+   <PropertyGroup>
+-    <TargetFramework>netcoreapp2.0</TargetFramework>
++    <TargetFramework>netcoreapp3.1</TargetFramework>
+   </PropertyGroup>
+ 
+   <ItemGroup>
+@@ -14,11 +14,12 @@
+   </ItemGroup>
+ 
+   <ItemGroup>
+-    <PackageReference Include="Microsoft.AspNetCore.All" Version="2.0.9" />
++    <DotNetCliToolReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Tools" Version="2.0.1" />
+   </ItemGroup>
+ 
+   <ItemGroup>
+-    <DotNetCliToolReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Tools" Version="2.0.1" />
++    <PackageReference Include="Microsoft.EntityFrameworkCore.Relational" Version="5.0.5" />
++    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="5.0.5" />
+   </ItemGroup>
+ 
+   <ItemGroup>
+diff --git a/SieveUnitTests/SieveUnitTests.csproj b/SieveUnitTests/SieveUnitTests.csproj
+index 772896a..afc9cf8 100644
+--- a/SieveUnitTests/SieveUnitTests.csproj
++++ b/SieveUnitTests/SieveUnitTests.csproj
+@@ -1,12 +1,13 @@
+ <Project Sdk="Microsoft.NET.Sdk">
+ 
+   <PropertyGroup>
+-    <TargetFramework>netcoreapp2.0</TargetFramework>
++    <TargetFramework>netcoreapp3.1</TargetFramework>
+ 
+     <IsPackable>false</IsPackable>
+   </PropertyGroup>
+ 
+   <ItemGroup>
++    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="3.1.14" />
+     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="15.5.0" />
+     <PackageReference Include="MSTest.TestAdapter" Version="1.2.0" />
+     <PackageReference Include="MSTest.TestFramework" Version="1.2.0" />
+-- 
+2.26.2.windows.1
+
