+++ Sieve.UnitTests/Abstractions/Entity/IBaseEntity.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace Sieve.UnitTests.Abstractions.Entity
+{
+    public interface IBaseEntity
+    {
+        int Id { get; set; }
+        DateTimeOffset DateCreated { get; set; }
+    }
+}
