+++ Sieve.UnitTests/Abstractions/Entity/IComment.cs
@@ -1,0 +1,7 @@
+﻿namespace Sieve.UnitTests.Abstractions.Entity
+{
+    public interface IComment: IBaseEntity
+    {
+        string Text { get; set; }
+    }
+}
