+++ Sieve.UnitTests/Services/SieveOptionsAccessor.cs
@@ -1,0 +1,18 @@
+﻿using Microsoft.Extensions.Options;
+using Sieve.Models;
+
+namespace Sieve.UnitTests.Services
+{
+    public class SieveOptionsAccessor : IOptions<SieveOptions>
+    {
+        public SieveOptions Value { get; }
+
+        public SieveOptionsAccessor()
+        {
+            Value = new SieveOptions
+            {
+                ThrowExceptions = true
+            };
+        }
+    }
+}
