--- SieveUnitTests/Abstractions/Entity/IBaseEntity.cs
@@ -1,10 +1,0 @@
-﻿using System;
-
-namespace SieveUnitTests.Abstractions.Entity
-{
-    public interface IBaseEntity
-    {
-        int Id { get; set; }
-        DateTimeOffset DateCreated { get; set; }
-    }
-}
