--- SieveUnitTests/Abstractions/Entity/IComment.cs
@@ -1,7 +1,0 @@
-﻿namespace SieveUnitTests.Abstractions.Entity
-{
-    public interface IComment: IBaseEntity
-    {
-        string Text { get; set; }
-    }
-}
