+++ azure-pipelines.yml
@@ -1,0 +1,30 @@
+# ------------------------------------------------------------------------------
+# <auto-generated>
+#
+#     This code was generated.
+#
+#     - To turn off auto-generation set:
+#
+#         [AzurePipelines (AutoGenerate = false)]
+#
+#     - To trigger manual generation invoke:
+#
+#         nuke --generate-configuration AzurePipelines --host AzurePipelines
+#
+# </auto-generated>
+# ------------------------------------------------------------------------------
+
+stages:
+  - stage: ubuntu_latest
+    displayName: 'ubuntu-latest'
+    dependsOn: [  ]
+    pool:
+      vmImage: 'ubuntu-latest'
+    jobs:
+      - job: Ci
+        displayName: 'Ci'
+        dependsOn: [  ]
+        steps:
+          - task: CmdLine@2
+            inputs:
+              script: './build.cmd Clean Restore Compile Test Package Ci --skip'
